One project page for every template: ProjectShell renders declared summaries
